Add score pruning
Add pruning functionality. Scores can now be set to expire on an interval. If a user does not have their score altered within the interval their score is deleted. This interval defaults to 2 weeks. Also adds a prune console command which enables the explicit deletion of a user's score or the explicit invocation of the pruneTable() function.
Also creates a new value in UserScore objects (UserScore.prototype.update) that contains the last update time of the score. When old score tables are loaded this value will be added and set to the current time.
Also adds interval validation. Valid intervals are between 0 and 2147483647 ms. Intervals outside this range are considered to mean disabling the interval all together.
The scoring command has also been updated to reflect that intervals may now be disabled